Update patch set 12
Patch Set 12:

If you remove the APLL_VAL_1200 switch for 1.2 GHz, then what is the point of any of the "overclocking" code?

Without the master APLL running at 1.2 GHz, then the CPU is running at 1.0 GHz, then all you have done here is reduce performance at the "OC" step by underclocking the buses.  OC0 is an underclock, not an overclock, so why is it labeled as an overclock?

Patch-set: 12