Update patch set 12
Patch Set 12:

No.  The clock divider takes care of that for 800 MHz and below.  See the first value in each entry of clkdiv_val

The formula is:
armclk = APLL/(DIVAPLL+1) (There may be an additional divisor in there, but that's usually set to 1, don't have the datasheet open)

I suggest hunting down the GalaxyS Hack Pack and looking at the CPU datasheet.  :)

Patch-set: 12