Call cpufreq_update_policy on DVFS events.
Calls cpufreq_update_policy on DVFS lock/unlock events to ensure that DVFS
locks take effect with the governor's preferred frequency (if higher than
the lock frequency).  Previously the lock would set the frequency to the
target level and expect the governor to reevaluate on a timer, which
ondemand does, but powersave and performance do not.

Also update the condition for which a PLL change is needed to prevent
lockups when changing max frequency (an issue not present prior to this patch).

Change-Id: I19a454d7173afdb99d0a87845285f84606cf24e1