Update patch set 1
Patch Set 1:

Correct, I added that small change to the patch (see the second paragraph of the commit message).  On epicmtd, without that change the device would lockup when changing the max frequency in the performance menu.  As I look at it now, it would appear to always be true using the current s5pv210_freq_table.  Perhaps the if can be removed, unless a more appropriate solution is discovered.

Patch-set: 1