Update patch set 2
Patch Set 2:

yeah, that's why I hadn't +1 or added any reviewers for this patch yet. Sure, we could add an a new service to aries-common to handle it , but I figured it would be better to fix it at the source as it would benefit other devices (aries are not the only devices that importing of init.local.rc does not work for).

Patch-set: 2