Update patch set 1
Patch Set 1: Looks good to me, but someone else must approve

I know that OC had previously been added and subsequently removed, but this patch does not appear to exhibit the same issues as the previous patch.

I've been using it for the past day or so with no ill effects, and regular usage + benchmark tools/etc confirm better performance, as well as the fact that the frequencies are updating as they should.

1.2GHz + interactive governor running for the past 24 hours with no reboots, freezes, crashes or other anomalies.

Patch-set: 1