Update patch set 1
Patch Set 1:

Steven:  If you did it in a way that redefined the enum for 800 MHz, you might have broken the suspend/resume code.  That's part of why this uses OC0 for 1.2 GHz instead of redefining L1-L4.

Biggest worry of mine would be the weird shit we saw on Infuse where adding the extra step would sometimes cause the CPU to get jammed at whatever the minimum frequency set in cpufreq policy was.  It was REALLY weird.

Patch-set: 1