Update patch set 1
Patch Set 1:

Andrew: The method we "attempted" on the epicmtd very much resembles how this one was handled on aries, with the exception LM1 instead OC0, on our part, trying to track down our sleep issue when running at a higher frequency however has been like running into a brick wall so far.

jt1134: Oh alright, I see. Mike(mkasick) had adopted an approach like this one when we were on 2.6.35 so not too different. Today for sake of testing I tried out this method to see how it'd work on the d700 and our issue of it going into a limbo state still persists, I'm glad however that it's not plaguing aries, but unfortunate that it's screwing with the d700.

Patch-set: 1