Update patch set 1
Patch Set 1:

I will fix the spacing issue no problem.

The reason I chose this freq_uv_table approach is that it is widely used and tested, and is the same approach as was used for UV in the CM7 kernel. I don't really see a benefit of creating a completely different approach.

Patch-set: 1