Update patch set 2
Patch Set 2: No score

Sorry for not explaining the -1
I am still looking for a easy way for the user to change the voltage, as all voltage control apps I tried only allow undervolting and not overvolting. I now manually change the sysfs to check whether it works, but there must be a better/easier way.

Patch-set: 2