Update patch set 2
Patch Set 2:

That'd be the problem of the app, we can raise the default voltage if that's needed (as long as it doesn't affect non-OC users). I want to keep this simple and don't want us to be in a position where we'll have dozens of patches tuning OC values (and affect stability for non-OC users).

Patch-set: 2