Update patch set 7
Patch Set 7:

thanks andrew, I completely forgot to compare with the infuse :)

Changing int_volt to 1125 and arm_volt to 1300 doesn't cause SoD for me. Previously setting arm_volt to anything higher than 1275 would cause SoD's.

Perhaps if the I9000 users that were having trouble with lower voltages are OK with 1300 + 1125 then we can just set it that way and leave it be.

Patch-set: 7