Update patch set 1
Patch Set 1:

It's done this way because Kconfig is from upstream, affects every device and architecture and should not be modified.

The current convention in crespo/aries defconfig is to only specify the values which override the default and thus these two values overriding the default are specified there.

Patch-set: 1