Update patch set 1
Patch Set 1:

Look, it's there in advanced hidden performance settings for those that want to use it. Normal people don't OC at all. If they want to OC, they better do their research.

The entire point is that this patch may work for you, but it can cause problems for other people.

Patch-set: 1