Update patch set 1
Patch Set 1:

Hmm, I the Huawei G300 has it. It has an option called 'cdrom_index' in /sys which sets a single lun-device to emulate a cdrom.

The device ships with autorun.iso, which is hosted through UMS with a cdrom-enabled lun. It allows Windows (or any other OS) to see the ISO as an CD-rom device.

The kernel is available here: http://download-c.huawei.com/tcpsdownload/downLoadCenter?category=&flay=software&downloadID=NDU5NTk=
It works a bit different, since you set cdrom_index which indirectly enables cdrom-support for the lun with that index.



Huawei creates a file here:

    /sys/devices/platform/s3c-usbgadget/gadget/cdrom_index

which can be written/read with a index:

    echo 0 > /sys/devices/platform/s3c-usbgadget/gadget/cdrom_index

Will enable cdrom-emulation for lun device with index 0 and disable cdrom-emulation for all other lun devices.

    echo 1 > /sys/devices/platform/s3c-usbgadget/gadget/cdrom_index

Will enable cdrom-emulation for lun device with index 1 and disable cdrom-emulation for all other lun devices.




My patch should be much more intuitive. It creates the files (one for each lun):

    /sys/devices/platform/s3c-usbgadget/gadget/lun0/cdrom
    /sys/devices/platform/s3c-usbgadget/gadget/lun1/cdrom

which can be written/read with a boolean (0 or 1):

    echo 1 > /sys/devices/platform/s3c-usbgadget/gadget/lun0/cdrom

enables cdrom-emulation for lun device with index 0.

    echo 0 > /sys/devices/platform/s3c-usbgadget/gadget/lun0/cdrom

disables cdrom-emulation for lun device with index 0.



CD-rom emulation is already available in the kernel. I just enable and disable it. It is disabled by default, so no-one should be harmed by this.

The other patch I submitted is to allow the emulated CD-rom to be detected correctly by the (PC-)bios too. It allows you to boot your PC from the (emulated) CD-rom device, which can hosts any ISO you choose (like a Windows install CD or Ubuntu LiveCD).

Patch-set: 1