Update patch set 1
Patch Set 1:

I'm the developer of DriveDroid ;). (Good to hear you know it!) DriveDroid is made to USB images. Most modern Linux ISOs are hybrid, which means they are specially crafted to work as USB-disk as well as CD-drive.
For non-hybrid ISOs this will not work. Examples of non-hybrid ISOs are all Windows-related ISOs, older Linux system and distributions where the developer is unaware of hybrid ISOs, etc. (outside the Linux world almost all ISOs are non-hybrid)

With this patch(+other patch) any ISO should be bootable.

I wouldn't worry about it not being used. Once there is support for the cdrom-option in DriveDroid (or any other application) there are many many people who want this. Some of them even want to buy a device specifically for this purpose. I'm not sure how viable it is to submit these patches to other device kernels of CyanogenMod, but I'd like this to be available for as many people as possible. If it ever gets that far people will be able to throw out their CD-drives, since those have become obsolete with this patch ;).

As for why it was not accepted in the first place, from what I've read on that mailing list CDrom-emulation first had support for 4096 blocksize, but because the specs described that you can choose 512 as well, the dev removed 4096 support: it cleaned up the code of some ifs (like the ones you see in the patch).
The problem however is that most bioses do not follow the specs and only allow 4096 blocksizes.

Another point in the discussion was to provide complete support for cdrom by also implementing CD-write support. I'm not sure what the practical use would be of that, but that's lacking in this patch.

Fact is, I haven't seen a better practical use for CDrom emulation than an Android phone. The code is all already in the kernel (except for the blocksize incompatiblity), it is just not enabled/controllable at runtime, which is what this patch enables.

Patch-set: 1