Update patch set 2
Patch Set 2: -Code-Review

Putting it in the Makefile is more in keeping with the kernel coding style, a quick "ag" of the kernel doesn't reveal any other place where a compiler warning is pragma'd out at the source. Also the maybe-uninitialized warning only seems to occur once. 

Personally I'd initialize it rather than risk raising an exception related to null values.

Really though they're all as bad as each other as no one seems to have a clear understanding ( me included ) of exactly what the code is doing, when it's doing it and why? 
So we're JFMIW ( Just F**king Making it Work )

meh!

Patch-set: 2