Update patch set 2
Patch Set 2:

> This exact code has been used since the ICS day. Only the warning is new ;)

GCC is getting a little more strict by default now-a-days and rightly so. I guess that it left Imagination in this state as I've got an Archos Honeycomb 2.6 kernel that has the exact same implementation. Who knows that null state could be totally intentional. It's probably overkill to search for the original commit and ask the dev about 1 line of code they wrote 3 years ago. It would be quite amusing though :)

> i am just removing -Werror from Makefile

Yeah go for it, fill your boots with that one.... :)

Patch-set: 2