Update patch set 2
Patch Set 2:

(1 comment)

Some of the ifdef were left intentionally because I doubt they will affect the other devices. For example the YP-G1 is the only device that have dual speakers. So leaving the necessary code for the right speaker won't do any harm to other Aries devices. I'm done with the ifdefs but I only wanted to test to see if it works nicely even with the corresponding code for YP-G1.

Patch-set: 2